+++ epicstel/__init__.py
@@ -1,3 +1,3 @@
 """EPICSTel main module"""
 
-__version__ = "4.0.1"
+__version__ = "4.1.0"