+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.0.1] - 2021-07-06
+## [4.1.0] - 2021-07-30
+### Added
+- Adds BBB state monitoring
+
+## [4.0.1] - 2021-07-21
 ### Changed
 - /unsubscribepv and /subscribepv functions are now able to affect group subscribes dynamically 
+- /checkme shows ignored PVs
 
 ## [4.0.0] - 2021-06-17
 ### Added